Move check-in payload functionality to passport-interface
