+++ apps/passport-server/src/services/generic-issuance/pipelines/LemonadePipeline.ts
@@ -8,7 +8,6 @@ import {
 import { EmailPCDPackage } from "@pcd/email-pcd";
 import { getHash } from "@pcd/passport-crypto";
 import {
-  GenericCheckinCredentialPayload,
   GenericIssuanceCheckInError,
   GenericIssuanceCheckInRequest,
   GenericIssuanceCheckInResponseValue,
@@ -22,14 +21,11 @@ import {
   PipelineType,
   PollFeedRequest,
   PollFeedResponseValue,
+  verifyCheckinCredential,
   verifyFeedCredential
 } from "@pcd/passport-interface";
 import { PCDActionType } from "@pcd/pcd-collection";
-import { ArgumentTypeName, SerializedPCD } from "@pcd/pcd-types";
-import {
-  SemaphoreSignaturePCD,
-  SemaphoreSignaturePCDPackage
-} from "@pcd/semaphore-signature-pcd";
+import { ArgumentTypeName } from "@pcd/pcd-types";
 import { str } from "@pcd/util";
 import { v5 as uuidv5 } from "uuid";
 import { LemonadeOAuthCredentials } from "../../../apis/lemonade/auth";
@@ -582,31 +578,6 @@ export class LemonadePipeline implements BasePipeline {
   }
 
   /**
-   * When checking tickets in, the user submits various pieces of data, wrapped
-   * in a Semaphore signature.
-   * Here we verify the signature, and return the encoded payload.
-   */
-  private async unwrapCheckInSignature(
-    credential: SerializedPCD<SemaphoreSignaturePCD>
-  ): Promise<GenericCheckinCredentialPayload> {
-    const signaturePCD = await SemaphoreSignaturePCDPackage.deserialize(
-      credential.pcd
-    );
-    const signaturePCDValid =
-      await SemaphoreSignaturePCDPackage.verify(signaturePCD);
-
-    if (!signaturePCDValid) {
-      throw new Error("Invalid signature");
-    }
-
-    const payload: GenericCheckinCredentialPayload = JSON.parse(
-      signaturePCD.claim.signedMessage
-    );
-
-    return payload;
-  }
-
-  /**
    * Given a ticket to check in, and a set of tickets belonging to the user
    * performing the check-in, verify that at least one of the user's tickets
    * belongs to a matching event and is a superuser ticket.
@@ -691,10 +662,9 @@ export class LemonadePipeline implements BasePipeline {
         let ticketId: string;
 
         try {
-          const payload = await this.unwrapCheckInSignature(request.credential);
-          const checkerEmailPCD = await EmailPCDPackage.deserialize(
-            payload.emailPCD.pcd
-          );
+          const payload = await verifyCheckinCredential(request.credential);
+          ticketId = payload.ticketIdToCheckIn;
+          const checkerEmailPCD = payload.emailPCD;
 
           if (
             !isEqualEdDSAPublicKey(
@@ -712,7 +682,6 @@ export class LemonadePipeline implements BasePipeline {
             this.id,
             checkerEmailPCD.claim.emailAddress
           );
-          ticketId = payload.ticketIdToCheckIn;
 
           span?.setAttribute("ticket_id", ticketId);
           span?.setAttribute(
@@ -724,6 +693,7 @@ export class LemonadePipeline implements BasePipeline {
             checkerEmailPCD.claim.semaphoreId
           );
         } catch (e) {
+          logger(`${LOG_TAG} Failed to verify credential due to error: `, e);
           setError(e, span);
           span?.setAttribute("precheck_error", "InvalidSignature");
           return { canCheckIn: false, error: { name: "InvalidSignature" } };
@@ -812,10 +782,9 @@ export class LemonadePipeline implements BasePipeline {
       let ticketId: string;
 
       try {
-        const payload = await this.unwrapCheckInSignature(request.credential);
-        const checkerEmailPCD = await EmailPCDPackage.deserialize(
-          payload.emailPCD.pcd
-        );
+        const payload = await verifyCheckinCredential(request.credential);
+        ticketId = payload.ticketIdToCheckIn;
+        const checkerEmailPCD = payload.emailPCD;
 
         if (
           !isEqualEdDSAPublicKey(
@@ -841,6 +810,7 @@ export class LemonadePipeline implements BasePipeline {
           checkerEmailPCD.claim.semaphoreId
         );
       } catch (e) {
+        logger(`${LOG_TAG} Failed to verify credential due to error: `, e);
         setError(e, span);
         span?.setAttribute("checkin_error", "InvalidSignature");
         return { checkedIn: false, error: { name: "InvalidSignature" } };
