+++ apps/passport-server/src/services/generic-issuance/pipelines/PretixPipeline.ts
@@ -9,7 +9,6 @@ import { EmailPCDPackage } from "@pcd/email-pcd";
 import { getHash } from "@pcd/passport-crypto";
 import {
   CSVPipelineDefinition,
-  GenericCheckinCredentialPayload,
   GenericIssuanceCheckInError,
   GenericIssuanceCheckInRequest,
   GenericIssuanceCheckInResponseValue,
@@ -30,14 +29,11 @@ import {
   PretixEventConfig,
   PretixPipelineDefinition,
   PretixProductConfig,
+  verifyCheckinCredential,
   verifyFeedCredential
 } from "@pcd/passport-interface";
 import { PCDActionType } from "@pcd/pcd-collection";
-import { ArgumentTypeName, SerializedPCD } from "@pcd/pcd-types";
-import {
-  SemaphoreSignaturePCD,
-  SemaphoreSignaturePCDPackage
-} from "@pcd/semaphore-signature-pcd";
+import { ArgumentTypeName } from "@pcd/pcd-types";
 import { normalizeEmail, str } from "@pcd/util";
 import { v5 as uuidv5 } from "uuid";
 import { IGenericPretixAPI } from "../../../apis/pretix/genericPretixAPI";
@@ -723,31 +719,6 @@ export class PretixPipeline implements BasePipeline {
   }
 
   /**
-   * When checking tickets in, the user submits various pieces of data, wrapped
-   * in a Semaphore signature.
-   * Here we verify the signature, and return the encoded payload.
-   */
-  private async unwrapCheckInSignature(
-    credential: SerializedPCD<SemaphoreSignaturePCD>
-  ): Promise<GenericCheckinCredentialPayload> {
-    const signaturePCD = await SemaphoreSignaturePCDPackage.deserialize(
-      credential.pcd
-    );
-    const signaturePCDValid =
-      await SemaphoreSignaturePCDPackage.verify(signaturePCD);
-
-    if (!signaturePCDValid) {
-      throw new Error("Invalid signature");
-    }
-
-    const payload: GenericCheckinCredentialPayload = JSON.parse(
-      signaturePCD.claim.signedMessage
-    );
-
-    return payload;
-  }
-
-  /**
    * Given a ticket to check in, and a set of tickets belonging to the user
    * performing the check-in, verify that at least one of the user's tickets
    * belongs to a matching event and is a superuser ticket.
@@ -819,10 +790,10 @@ export class PretixPipeline implements BasePipeline {
         let ticketId: string;
 
         try {
-          const payload = await this.unwrapCheckInSignature(request.credential);
-          const checkerEmailPCD = await EmailPCDPackage.deserialize(
-            payload.emailPCD.pcd
-          );
+          const payload = await verifyCheckinCredential(request.credential);
+          ticketId = payload.ticketIdToCheckIn;
+          const checkerEmailPCD = payload.emailPCD;
+
           if (
             !isEqualEdDSAPublicKey(
               checkerEmailPCD.proof.eddsaPCD.claim.publicKey,
@@ -939,10 +910,9 @@ export class PretixPipeline implements BasePipeline {
       let ticketId: string;
 
       try {
-        const payload = await this.unwrapCheckInSignature(request.credential);
-        const checkerEmailPCD = await EmailPCDPackage.deserialize(
-          payload.emailPCD.pcd
-        );
+        const payload = await verifyCheckinCredential(request.credential);
+        ticketId = payload.ticketIdToCheckIn;
+        const checkerEmailPCD = payload.emailPCD;
 
         if (
           !isEqualEdDSAPublicKey(
